% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr.R
\docType{data}
\name{allCrime}
\alias{allCrime}
\title{CrimeTypes}
\format{An object of class \code{character} of length 56.}
\usage{
allCrime
}
\arguments{
\item{allCrime}{contains crime categories}
}
\description{
CrimeTypes
}
\keyword{datasets}
