% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr.R
\name{getCrimeByID}
\alias{getCrimeByID}
\title{Function to mao crime name with Id}
\usage{
getCrimeByID(id)
}
\arguments{
\item{id}{crime ID associated with crime}
}
\description{
Function to mao crime name with Id
}
