% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr.R
\name{getLimitedData}
\alias{getLimitedData}
\title{Get data limted by number of observations}
\usage{
getLimitedData(limit = 5000)
}
\arguments{
\item{limit}{to limit records
limit the records}
}
\value{
dataset containing limited rows with limit as \code{limit}
}
\description{
Get data limted by number of observations
}
\examples{
getLimitedData(limit = 5000)

}
