% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr.R
\docType{data}
\name{baseUrl}
\alias{baseUrl}
\title{Base Url to get crime data}
\format{An object of class \code{character} of length 1.}
\usage{
baseUrl
}
\description{
Base Url to get crime data
}
\keyword{datasets}
