% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr.R
\name{getDataByCrimeType}
\alias{getDataByCrimeType}
\title{Function to get data by crime type}
\usage{
getDataByCrimeType(dataset, crimeType)
}
\arguments{
\item{dataset}{a dataset}

\item{crimeType}{a crime type}
}
\value{
dataset for a specific crime given by \code{crimeType}
}
\description{
Function to get data by crime type
}
